Iterator cleanups (using ptrdiff_t instead of int in a couple places) and progress on RandomAccessIterator

+++ Library/Sources/Stroika/Foundation/Traversal/BidirectionalIterator.h
@@ -81,7 +81,7 @@ namespace Stroika::Foundation::Traversal {
          *
          *  \pre i >= 0
          */
-        nonvirtual BidirectionalIterator operator- (int i) const;
+        nonvirtual BidirectionalIterator operator- (ptrdiff_t i) const;
 
     public:
         /**
